app/requirements.txt
flask==3.1.0
pyyaml==6.0.2
netmiko==4.5.0
werkzeug==3.1.3
gunicorn==23.0.0
